Updated: Moved directive to its own file
